move button in a separated componet
